pyproject.toml
[tool.poetry]
name = "python-zenmoney"
version = "0.1.0"
description = ""
authors = ["VaryaKasha <varyaxx@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.5"
requests = "^2.23.0"
pyTelegramBotAPI = "^3.6.7"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
